Update with latest papers.csv and simplify code
