+++ scripts/requirements.txt
@@ -1,5 +1,6 @@
 transformers
 sklearn
 umap-learn
+openreview-py
 torch==1.4.0
 ics
